tests/wpt/tests/html/canvas/element/layers/2d.layer.ctm.getTransform.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.layer.ctm.getTransform</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.layer.ctm.getTransform</h1>
<p class="desc">Tests getTransform inside layers.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
test(t => {
  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  ctx.translate(10, 20);
  ctx.beginLayer();
  ctx.scale(2, 3);
  const m = ctx.getTransform();
  assert_array_equals([m.a, m.b, m.c, m.d, m.e, m.f], [2, 0, 0, 3, 10, 20]);
  ctx.endLayer();

}, "Tests getTransform inside layers.");
</script>
